# docs/contracts/openapi/backend-bff-v1.json
{
  "components": {
    "schemas": {
...
      "OperationStatusResponse": {
        "additionalProperties": false,
        "properties": {
          "attempt": {
            "minimum": 0.0,
            "title": "Attempt",
            "type": "integer"
          },
          "correlation_id": {
            "title": "Correlation Id",
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "error_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Code"
          },
          "operation_id": {
            "title": "Operation Id",
            "type": "string"
          },
          "operation_type": {
            "title": "Operation Type",
            "type": "string"
          },
          "result_resource_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Result Resource Id"
          },
          "schema_version": {
            "const": "operation_status_response.v1",
            "default": "operation_status_response.v1",
            "title": "Schema Version",
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/PublicOperationStatus"
          },
          "subject_id": {
            "title": "Subject Id",
            "type": "string"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "operation_id",
          "operation_type",
          "subject_id",
          "status",
          "attempt",
          "result_resource_id",
          "error_code",
          "correlation_id",
          "created_at",
          "updated_at"
        ],
        "title": "OperationStatusResponse",
        "type": "object"
      },
...
      "PublicErrorCode": {
        "enum": [
          "AUTHENTICATION_REQUIRED",
          "INVALID_SERVICE_CREDENTIAL",
          "INVALID_ACTOR_ASSERTION",
          "ACTOR_ASSERTION_REPLAYED",
          "AUTHORIZATION_DENIED",
          "AUTHORIZATION_UNAVAILABLE",
          "RESOURCE_NOT_FOUND",
          "DATA_INSUFFICIENT",
          "RESULT_PENDING",
          "IDEMPOTENCY_CONFLICT",
          "INVALID_REQUEST",
          "OPERATION_CONFLICT",
          "OPERATION_LEASE_LOST",
          "BINDING_REQUIRED",
          "INVALID_LIFECYCLE_TRANSITION",
          "PROVIDER_UNAVAILABLE",
          "CURSOR_RESYNC_REQUIRED",
          "INTERNAL_ERROR"
        ],
        "title": "PublicErrorCode",
        "type": "string"
      },
...
      "PublicOperationStatus": {
        "enum": [
          "pending",
          "running",
          "succeeded",
          "failed",
          "cancelled"
        ],
        "title": "PublicOperationStatus",
        "type": "string"
      },
...
      "RevocationTombstone": {
        "additionalProperties": false,
        "properties": {
          "action_required": {
            "const": "delete_local_authorized_cache",
            "default": "delete_local_authorized_cache",
            "title": "Action Required",
            "type": "string"
          },
          "aggregate_id": {
            "title": "Aggregate Id",
            "type": "string"
          },
          "aggregate_version": {
            "const": 1,
            "default": 1,
            "title": "Aggregate Version",
            "type": "integer"
          },
          "effective_at": {
            "format": "date-time",
            "title": "Effective At",
            "type": "string"
          },
          "event_id": {
            "title": "Event Id",
            "type": "string"
          },
          "event_type": {
            "const": "ACCESS_REVOKED_TOMBSTONE",
            "default": "ACCESS_REVOKED_TOMBSTONE",
            "title": "Event Type",
            "type": "string"
          },
          "per_aggregate_sequence": {
            "const": 1,
            "default": 1,
            "title": "Per Aggregate Sequence",
            "type": "integer"
          },
          "remote_recall_possible": {
            "const": false,
            "default": false,
            "title": "Remote Recall Possible",
            "type": "boolean"
          },
          "schema_version": {
            "const": "revocation_tombstone.v1",
            "default": "revocation_tombstone.v1",
            "title": "Schema Version",
            "type": "string"
          },
          "subject_id": {
            "title": "Subject Id",
            "type": "string"
          }
        },
        "required": [
          "event_id",
          "aggregate_id",
          "subject_id",
          "effective_at"
        ],
        "title": "RevocationTombstone",
        "type": "object"
      },
...
      "sleepagent__api__public_contracts__ErrorResponse": {
        "additionalProperties": false,
        "properties": {
          "code": {
            "$ref": "#/components/schemas/PublicErrorCode"
          },
          "correlation_id": {
            "maxLength": 128,
            "minLength": 1,
            "title": "Correlation Id",
            "type": "string"
          },
          "details": {
            "additionalProperties": {
              "type": "string"
            },
            "title": "Details",
            "type": "object"
          },
          "message": {
            "maxLength": 300,
            "minLength": 1,
            "title": "Message",
            "type": "string"
          },
          "retryable": {
            "default": false,
            "title": "Retryable",
            "type": "boolean"
          },
          "schema_version": {
            "const": "error.v1",
            "default": "error.v1",
            "title": "Schema Version",
            "type": "string"
          },
          "tombstone": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/RevocationTombstone"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "required": [
          "code",
          "message",
          "correlation_id"
        ],
        "title": "ErrorResponse",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "SleepActorAssertion": {
        "description": "Asymmetric JWS assertion bound to this HTTP request.",
        "in": "header",
        "name": "X-Sleep-Actor-Assertion",
        "type": "apiKey"
      },
      "SleepServiceCredential": {
        "description": "Rotating HTTPS Bearer credential for the calling service.",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "SleepAgent Backend",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/v1/operations/{operation_id}": {
      "get": {
        "operationId": "getSleepOperationV1",
        "parameters": [
          {
            "in": "path",
            "name": "operation_id",
            "required": true,
            "schema": {
              "title": "Operation Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OperationStatusResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/sleepagent__api__public_contracts__ErrorResponse"
                }
              }
            },
            "description": "Bad Request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/sleepagent__api__public_contracts__ErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/sleepagent__api__public_contracts__ErrorResponse"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/sleepagent__api__public_contracts__ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/sleepagent__api__public_contracts__ErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/sleepagent__api__public_contracts__ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Content"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/sleepagent__api__public_contracts__ErrorResponse"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "security": [
          {
            "SleepServiceCredential": []
          },
          {
            "SleepActorAssertion": []
          }
        ],
        "summary": "Get Operation",
        "tags": [
          "Sleep domain v1"
        ]
      }
    },
...
  }
}